Added macros for precision and other values; fixed the algorithm. The code is now prod-ready.
Added copyright notice in README.
Added help message when no arguments are passed into the program.
Changed the algorithm to catch when our degree is 0 or 1.
Changed the algorithm where it now fuCKING WORKS.
Added macros to determine whether to use fixed notation, number of digits
for precision, etc.
Added number of microseconds used to perform the algorithm in main().

Made with tender-loving care.